+++ game/shared/fxbutemgr.cpp
@@ -4217,7 +4217,7 @@ CSpecialFX* CFXButeMgr::CreateScaleFX(CScaleFX* pScaleFX, const LTVector &vPos,
 	scale.fDelayTime		= pScaleFX->fDelayTime;
 	scale.bAdditive			= pScaleFX->bAdditive;
 	scale.bMultiply			= pScaleFX->bMultiply;
-	scale.nType				= (pScaleFX->eType == SCALEFX_MODEL) ? OT_MODEL : OT_SPRITE;
+	scale.nType				= static_cast<uint8>((pScaleFX->eType == SCALEFX_MODEL) ? OT_MODEL : OT_SPRITE);
     scale.bUseUserColors    = LTTRUE;
 	scale.bFaceCamera		= pScaleFX->bFaceCamera;
 	scale.nRotationAxis		= pScaleFX->nRotationAxis;
@@ -4268,7 +4268,7 @@ CSpecialFX* CFXButeMgr::CreatePShowerFX(CPShowerFX* pPShowerFX, const LTVector &
 	ps.vColor1			= pPShowerFX->vColor1;
 	ps.vColor2			= pPShowerFX->vColor2;
 	ps.pTexture			= pPShowerFX->szTexture;
-	ps.nParticles		= GetRandom(pPShowerFX->nMinParticles, pPShowerFX->nMaxParticles);
+	ps.nParticles		= static_cast<uint8>(GetRandom(pPShowerFX->nMinParticles, pPShowerFX->nMaxParticles));
 	ps.fDuration		= GetRandom(pPShowerFX->fMinDuration, pPShowerFX->fMaxDuration);
 	ps.fEmissionRadius	= pPShowerFX->fEmissionRadius;
 	ps.fRadius			= pPShowerFX->fRadius;
@@ -4397,7 +4397,7 @@ CSpecialFX* CFXButeMgr::CreatePExplFX(PEXPLFX* pPExplFX, const LTRotation &rSurf
 	PESCREATESTRUCT pe;
     pe.rSurfaceRot = rSurfaceRot;
 
-	pe.nSurfaceType		= eSurfaceType;
+	pe.nSurfaceType		= static_cast<uint8>(eSurfaceType);
 	pe.vPos				= vPos + pPExplFX->vPosOffset;
 	pe.bCreateDebris	= pPExplFX->bCreateDebris;
 	pe.bRotateDebris	= pPExplFX->bRotateDebris;
@@ -4413,9 +4413,9 @@ CSpecialFX* CFXButeMgr::CreatePExplFX(PEXPLFX* pPExplFX, const LTRotation &rSurf
 	pe.fOffsetTime		= pPExplFX->fOffsetTime;
 	pe.fRadius			= pPExplFX->fRadius;
 	pe.fGravity			= pPExplFX->fGravity;
-	pe.nNumPerPuff		= pPExplFX->nNumPerPuff;
-	pe.nNumEmitters		= pPExplFX->nNumEmitters;
-	pe.nNumSteps		= pPExplFX->nNumSteps;
+	pe.nNumPerPuff		= static_cast<uint8>(pPExplFX->nNumPerPuff);
+	pe.nNumEmitters		= static_cast<uint8>(pPExplFX->nNumEmitters);
+	pe.nNumSteps		= static_cast<uint8>(pPExplFX->nNumSteps);
 	pe.pFilename		= pPExplFX->szFile;
 	pe.bAdditive		= pPExplFX->bAdditive;
 	pe.bMultiply		= pPExplFX->bMultiply;
